fix tab-bar initial status bug
