+++ packages/theme-default/src/tabs.css
@@ -14,7 +14,6 @@
       position: absolute;
       bottom: 0;
       left: 0;
-      width: 100%;
       height: 3px;
       background-color: var(--color-primary);
       z-index: 1;
